Adding in event tests

+++ src/mantle/framework/providers/class-queue-service-provider.php
@@ -15,6 +15,7 @@ use Mantle\Framework\Queue\Events\Job_Processed;
 use Mantle\Framework\Queue\Events\Run_Complete;
 use Mantle\Framework\Queue\Queue_Manager;
 use Mantle\Framework\Queue\Worker;
+use Mantle\Framework\Queue\Wp_Cron_Job;
 use Mantle\Framework\Queue\Wp_Cron_Provider;
 use Mantle\Framework\Queue\Wp_Cron_Scheduler;
 use Mantle\Framework\Service_Provider;
@@ -87,10 +88,12 @@ class Queue_Service_Provider extends Service_Provider {
 		$this->app['events']->listen(
 			Job_Processed::class,
 			function( Job_Processed $event ) {
-				$queue_post_id = $event->job->get_post_id();
+				if ( $event->job instanceof Wp_Cron_Job ) {
+					$queue_post_id = $event->job->get_post_id();
 
-				if ( $queue_post_id ) {
-					wp_delete_post( $queue_post_id, true );
+					if ( $queue_post_id ) {
+						wp_delete_post( $queue_post_id, true );
+					}
 				}
 			}
 		);
@@ -99,7 +102,9 @@ class Queue_Service_Provider extends Service_Provider {
 		$this->app['events']->listen(
 			Run_Complete::class,
 			function( Run_Complete $event ) {
-				Wp_Cron_Scheduler::schedule_next_run( $event->queue );
+				if ( $event->provider instanceof Wp_Cron_Provider ) {
+					Wp_Cron_Scheduler::schedule_next_run( $event->queue );
+				}
 			}
 		);
 	}
